Fix displaying of Finna subpanel in Business Relationship's Detail View
SugarCRM's code is adding a filler columns to the last row of the subpanel
consisting of more than a single row (as in this case). Due to the
way Finna fields are implemented, they form a single column per row,
and a filler column only ruins the display. This adds a custom detail
view handler for Business Relationship that handles this special case.